% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlc_methods.R
\name{hlc_methods}
\alias{hlc_methods}
\title{List available HLC methods}
\usage{
hlc_methods()
}
\value{
A character vector of available method names.
}
\description{
Returns the currently implemented Herd Lying Concordance dispersion bases.
}
