% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlc_weighted_methods.R
\name{hlc_weighted_methods}
\alias{hlc_weighted_methods}
\title{List available lying-weighted HLC methods}
\usage{
hlc_weighted_methods()
}
\value{
A character vector of available lying-weighted HLC method names.
}
\description{
Returns the currently implemented lying-weighted Herd Lying Concordance
method names.
}
