% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlc_weighted_methods.R
\name{hlc_weighted_methods}
\alias{hlc_weighted_methods}
\title{List available lying-weighted HLC methods}
\usage{
hlc_weighted_methods()
}
\value{
A character vector containing the available lying-weighted HLC
method names.
}
\description{
Returns the names of the currently implemented lying-weighted Herd Lying
Concordance methods.
}
\examples{
hlc_weighted_methods()

}
